# backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:8081",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/estado/tipoEstado": {
      "get": {
        "tags": [
          "Estado"
        ],
        "description": "Endpoint para obtener los tipos de estado.",
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/estados"
            },
            "description": "estados."
          },
          "404": {
            "description": "Error "
          }
        }
      }
    },
    "/estado/{id}": {
      "get": {
        "tags": [
          "Estado"
        ],
        "description": "Endpoint para obtener las tareas asigandas de un operario.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID de operario."
          }
        ],
        "responses": {
          "200": {
            "description": "tareas del operario.",
            "schema": {
              "$ref": "#/definitions/tareas"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Error "
          }
        }
      }
    },
    "/estado/": {
      "put": {
        "tags": [
          "Estado"
        ],
        "description": "Endpoint para modificar el estado.",
        "parameters": [
          {
            "name": "cambioEstado",
            "in": "body",
            "description": "datos para modificar el estado de una tarea, esto puede generar el cambio de esatdo de la oti y de sector",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/cambioEstado"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "se cambio de estado"
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
...
    "/orden/": {
      "get": {
        "tags": [
          "Orden"
        ],
        "description": "Endpoint para obtener todas las ordenes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "ordenes encontradas",
            "schema": {
              "$ref": "#/definitions/Ordenes"
            }
          },
          "404": {
            "description": "Error "
          }
        }
      }
    },
    "/orden/{id}": {
      "get": {
        "tags": [
          "Orden"
        ],
        "description": "Endpoint para obtener una orden.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "orden ID"
          }
        ],
        "responses": {
          "200": {
            "description": "orden encontrada",
            "schema": {
              "$ref": "#/definitions/orden"
            }
          },
          "404": {
            "description": "Error "
          }
        }
      },
      "patch": {
        "tags": [
          "Orden"
        ],
        "description": "Endpoint para modificar orden.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "orden ID"
          },
          {
            "name": "orden",
            "in": "body",
            "description": "datos para modificar una orden",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/orden"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "se modifica orden",
            "schema": {
              "$ref": "#/definitions/orden"
            }
          },
          "404": {
            "description": "Error "
          }
        }
      }
    },
    "/orden/crearOrdenProduccion": {
      "post": {
        "tags": [
          "Orden"
        ],
        "description": "Endpoint para crear orden de produccion.",
        "parameters": [
          {
            "name": "orden",
            "in": "body",
            "description": "datos para crear una orden",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/orden"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "se crea orden",
            "schema": {
              "$ref": "#/definitions/orden"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Error "
          }
        }
      }
    },
...
    "/cliente/": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "description": "Endpoint para obtener todos los clientes",
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/clientes"
            },
            "description": "clientes"
          },
          "404": {
            "description": "Error "
          }
        }
      },
      "post": {
        "tags": [
          "Cliente"
        ],
        "description": "Endpoint para crear un cliente",
        "parameters": [
          {
            "name": "cliente",
            "in": "body",
            "description": "datos para crear cliente",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/cliente"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/cliente"
            },
            "description": "cliente"
          },
          "404": {
            "description": "Error "
          }
        }
      }
    },
    "/cliente/{cuit}": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "description": "Endpoint para obtener un cliente",
        "parameters": [
          {
            "name": "cuit",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "cuit de la empresa."
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/cliente"
            },
            "description": "cliente"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Error "
          }
        }
      }
    },
    "/cliente/{id}": {
      "put": {
        "tags": [
          "Cliente"
        ],
        "description": "Endpoint para modificar cliente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID cliente"
          },
          {
            "name": "cliente",
            "in": "body",
            "description": "datos para crear cliente",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/cliente"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Error "
          }
        }
      }
    },
    "/operario/": {
      "get": {
        "tags": [
          "Operario"
        ],
        "description": "Endpoint para obtener todos los operarios.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operarios encontrado",
            "schema": {
              "$ref": "#/definitions/operarios"
            }
          },
          "404": {
            "description": "Error "
          }
        }
      }
    },
    "/operario/{id}": {
      "get": {
        "tags": [
          "Operario"
        ],
        "description": "Endpoint para obtener las tareas.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Operario ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Operario encontrado",
            "schema": {
              "$ref": "#/definitions/operario"
            }
          },
          "404": {
            "description": "Error "
          }
        }
      }
    },
    "/operario/operarioSector/{nombreSector}": {
      "get": {
        "tags": [
          "Operario"
        ],
        "description": "Endpoint para obtener los operarios de un sector.",
        "parameters": [
          {
            "name": "nombreSector",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operarios sector encontrado",
            "schema": {
              "$ref": "#/definitions/operarios"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Error "
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Ordenes": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "619c3f28a5e105f51b94f683"
          },
          "detalle": {
            "type": "string",
            "example": "detalle"
          },
          "fecha": {
            "type": "string",
            "example": "2021-11-22"
          },
          "supervisor": {
            "type": "object",
            "properties": {
              "numeroEmpleado": {
                "type": "string",
                "example": "5"
              },
              "DNI": {
                "type": "string",
                "example": "205967118"
              },
              "nombre": {
                "type": "string",
                "example": "Suzanne"
              },
              "apellido": {
                "type": "string",
                "example": "Kleynermans"
              },
              "telefono": {
                "type": "string",
                "example": "253-396-2243"
              },
              "nombreUsuario": {
                "type": "string",
                "example": "skleynermans4"
              },
              "contraseña": {
                "type": "string",
                "example": "LACiCk"
              },
              "email": {
                "type": "string",
                "example": "skleynermans4@acquirethisname.com"
              },
              "_id": {
                "type": "string",
                "example": "613bc8c397d979667ca2a13a"
              }
            }
          },
          "oti": {
            "type": "array",
            "example": [],
            "items": {}
          },
          "rosca": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "descripcionTecnica": {
                  "type": "string",
                  "example": "descripcion"
                },
                "medida": {
                  "type": "string",
                  "example": "30x30x30"
                },
                "tipoDeRosca": {
                  "type": "object",
                  "properties": {
                    "descripcion": {
                      "type": "string",
                      "example": "descripcion_5"
                    },
                    "nombre": {
                      "type": "string",
                      "example": "TDL"
                    },
                    "categoria": {
                      "type": "object",
                      "properties": {
                        "nombre": {
                          "type": "string",
                          "example": "premium"
                        },
                        "descripcion": {
                          "type": "string",
                          "example": "exportacion marina y utilizacion para caños de gas"
                        },
                        "_id": {
                          "type": "string",
                          "example": "619c3f28a5e105f51b94f67e"
                        }
                      }
                    },
                    "_id": {
                      "type": "string",
                      "example": "613e604acb4275f300506790"
                    }
                  }
                },
                "_id": {
                  "type": "string",
                  "example": "619c3f28a5e105f51b94f67f"
                }
              }
            }
          }
        }
      }
    },
    "orden": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "619c3f28a5e105f51b94f683"
        },
        "detalle": {
          "type": "string",
          "example": "detalle"
        },
        "fecha": {
          "type": "string",
          "example": "2021-11-22"
        },
        "supervisor": {
          "type": "object",
          "properties": {
            "numeroEmpleado": {
              "type": "string",
              "example": "5"
            },
            "DNI": {
              "type": "string",
              "example": "205967118"
            },
            "nombre": {
              "type": "string",
              "example": "Suzanne"
            },
            "apellido": {
              "type": "string",
              "example": "Kleynermans"
            },
            "telefono": {
              "type": "string",
              "example": "253-396-2243"
            },
            "nombreUsuario": {
              "type": "string",
              "example": "skleynermans4"
            },
            "contraseña": {
              "type": "string",
              "example": "LACiCk"
            },
            "email": {
              "type": "string",
              "example": "skleynermans4@acquirethisname.com"
            },
            "_id": {
              "type": "string",
              "example": "613bc8c397d979667ca2a13a"
            }
          }
        },
        "oti": {
          "type": "array",
          "example": [],
          "items": {}
        },
        "rosca": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "descripcionTecnica": {
                "type": "string",
                "example": "descripcion"
              },
              "medida": {
                "type": "string",
                "example": "30x30x30"
              },
              "tipoDeRosca": {
                "type": "object",
                "properties": {
                  "descripcion": {
                    "type": "string",
                    "example": "descripcion_5"
                  },
                  "nombre": {
                    "type": "string",
                    "example": "TDL"
                  },
                  "categoria": {
                    "type": "object",
                    "properties": {
                      "nombre": {
                        "type": "string",
                        "example": "premium"
                      },
                      "descripcion": {
                        "type": "string",
                        "example": "exportacion marina y utilizacion para caños de gas"
                      },
                      "_id": {
                        "type": "string",
                        "example": "619c3f28a5e105f51b94f67e"
                      }
                    }
                  },
                  "_id": {
                    "type": "string",
                    "example": "613e604acb4275f300506790"
                  }
                }
              },
              "_id": {
                "type": "string",
                "example": "619c3f28a5e105f51b94f67f"
              }
            }
          }
        }
      }
    },
...
    "clientes": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "613e54dacb4275f300506776"
          },
          "DNI": {
            "type": "string",
            "example": "30698993.0"
          },
          "CUIT": {
            "type": "string",
            "example": "20306989934.0"
          },
          "nombre": {
            "type": "string",
            "example": "Melamie"
          },
          "apellido": {
            "type": "string",
            "example": "Bamblett"
          },
          "telefono": {
            "type": "string",
            "example": "724 673 2004"
          },
          "correo": {
            "type": "string",
            "example": "mbamblett3@wordpress.com"
          },
          "nombreEmpresa": {
            "type": "string",
            "example": "Vimbo"
          },
          "ordenProduccion": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "idOrdenProduccion": {
                  "type": "string",
                  "example": "613bc8c397d979667ddda137"
                },
                "fecha": {
                  "type": "string",
                  "example": "07/05/2021"
                },
                "detalle": {
                  "type": "string",
                  "example": "400 roscas para plataforma marina"
                },
                "estado": {
                  "type": "string",
                  "example": "finaliza"
                }
              }
            }
          }
        }
      }
    },
...
    "estados": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "613bcf1597d979667ca2a162"
          },
          "nombre": {
            "type": "string",
            "example": "finalizada"
          },
          "descripcion": {
            "type": "string",
            "example": "ha finalizado con exito"
          }
        }
      }
    },
    "tareas": {
      "type": "object",
      "properties": {
        "tareasOperario": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "descripcion": {
                "type": "string",
                "example": "asignar a operario con experiencia"
              },
              "nombre": {
                "type": "string",
                "example": "cortar pieza"
              },
              "fechaInicio": {
                "type": "string",
                "example": "YYYY/MM/DD"
              },
              "fechaFin": {
                "type": "string",
                "example": "YYYY/MM/DD"
              },
              "numeroDeOrden": {
                "type": "string",
                "example": "4589"
              },
              "tipoDeTarea": {
                "type": "string",
                "example": "tareas de corte"
              },
              "sector": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "corte"
                  },
                  "activo": {
                    "type": "string",
                    "example": "true"
                  },
                  "operario": {
                    "type": "array",
                    "example": [
                      "5ea7029320e515102e44c71d",
                      "5ea7029320e515102ed5c22d",
                      "5ea7029320e515222102ed5c71d",
                      "5ea7029320e515123pc71d"
                    ],
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "idOperario": {
                "type": "string",
                "example": "5ea7029320e515102e44c71d"
              },
              "estado": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "fechaInicio": {
                      "type": "string",
                      "example": "YYYY/MM/DD"
                    },
                    "fechaFin": {
                      "type": "string",
                      "example": "YYYY/MM/DD"
                    },
                    "observacion": {
                      "type": "string",
                      "example": "si inicia con todos las herramientas calibradas"
                    },
                    "tipoEstado": {
                      "type": "object",
                      "properties": {
                        "nombre": {
                          "type": "string",
                          "example": "en progreso"
                        },
                        "descripcion": {
                          "type": "string",
                          "example": "en proceso de produccion"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "idOti": {
          "type": "string",
          "example": "616f2980c5e1d1846975c8cd"
        }
      }
    },
    "cambioEstado": {
      "type": "object",
      "properties": {
        "idOti": {
          "type": "string",
          "example": "616f2980c5e1d1846975c8cd"
        },
        "idTarea": {
          "type": "string",
          "example": "616f2981c5e1d1846975c8f5"
        },
        "nombreSector": {
          "type": "string",
          "example": "corte"
        },
        "observacion": {
          "type": "string",
          "example": "se inicia proceso de produccion de tarea"
        },
        "tipoEstado": {
          "type": "object",
          "properties": {
            "nombre": {
              "type": "string",
              "example": "en progreso"
            },
            "descripcion": {
              "type": "string",
              "example": "en proceso de produccion"
            }
          }
        }
      }
    }
  }
}